+++ include/nstd/alloc/heap.h
@@ -40,14 +40,12 @@ NSTDAPI NSTDHeap nstd_alloc_heap_from_raw(const NSTDAny ptr, const NSTDUSize siz
 ///
 /// # Parameters
 ///
-/// - `const NSTDAny ptr` - A raw pointer to the heap object.
-///
-/// - `const NSTDUSize size` - The size of the heap object.
+/// - `const NSTDPointer *const ptr` - A pointer to the heap object.
 ///
 /// # Returns
 ///
 /// `NSTDHeap obj` - The new heap object.
-NSTDAPI NSTDHeap nstd_alloc_heap_from_existing(const NSTDAny ptr, const NSTDUSize size);
+NSTDAPI NSTDHeap nstd_alloc_heap_from_existing(const NSTDPointer *const ptr);
 
 /// Frees a heap allocated object.
 ///